Added "Teaz Intelligence"
ChatGPT can now be directly adressed by the user by putting "Teaz" in front of a message. This will soon also link to Teaz Actions, a format in the reply that lets chatGPT adress actions to execute in discord. Removed the image analysis functionality to later be readded into the "Teaz Intelligence" feature.